Add files via upload
MovieRental Website
This web page is made in reactjs. It's a dynamic website. The site helps the user to watch and rent movies. This is just a demo website hence none of the cards redirects to any dedicated film of its own also the buttons are made static. The web page is broken down into different components i.e the header, main content pages, and last but not the least footer. The input bar present is dynamic i.e. it takes you to the particular searched content of any type in seconds.